\name{farmer3}
\alias{farmer3}
\docType{data}
\title{
Kansas Farmer Data of Bilder and Loughin (2007)
}
\description{
Survey of 279 Kansas farmers conducted by the Department of Animal Sciences at Kansas State University.  Data are provided for only three of the survey questions.
}
\usage{farmer3}
\format{
The data frame contains 279 rows and the following 12 columns:

Columns 1-3 correspond to the response categories for the question "Which of the following do you test your swine waste for?" Binary responses (1 = Yes, 0 = No) are provided for each category.
\itemize{
\item{\code{W1}} {Nitrogen}
\item{\code{W2}} {Phosphorous}
\item{\code{W3}} {Salt}
}
Columns 4-7 correspond to the response categories for the question "What swine waste disposal methods do you use?" Binary responses (1 = Yes, 0 = No) are provided for each category.
\itemize{
\item{\code{Y1}} {Lagoon}
\item{\code{Y2}} {Pit}
\item{\code{Y3}} {Natural drainage}
\item{\code{Y4}} {Holding tank}
}
Columns 8-12 correspond to the response categories for the question "What are your primary sources of veterinary information?" Binary responses (1 = Yes, 0 = No) are provided for each category.
\itemize{
\item{\code{Z1}} {Professional consultant}
\item{\code{Z2}} {Veterinarian}
\item{\code{Z3}} {State or local extension service}
\item{\code{Z4}} {Magazines}
\item{\code{Z5}} {Feed companies and representatives}
}
}
\source{
Bilder, C. and Loughin, T. (2007)  Modeling association between two or more categorical variables that allow for multiple category choices.  \emph{Communications in Statistics--Theory and Methods}, \bold{36}, 433--451.
}