\name{print.MMI}
\alias{print.MMI}
\title{
Control Printed Display of Objects of Class "MMI"
}
\description{
The \code{print.MMI} method function controls the printed display of objects of class \code{MMI}.
}
\usage{
\method{print}{MMI}(x, ...)
}
\arguments{
  \item{x}{
An object of class \code{"MMI"} produced by the \code{\link{MI.test}} function.
}
  \item{...}{
Additional arguments passed to or from other methods.
}
}
\value{
A print out of selected results from \code{\link{MI.test}}.
}