% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.MMI.R, R/print.SPMI.R,
%   R/print.anova.genloglin.R, R/print.genloglin.R, R/print.predict.genloglin.R,
%   R/print.summary.genloglin.R
\name{print.MMI}
\alias{print.MMI}
\alias{print.SPMI}
\alias{print.anova.genloglin}
\alias{print.genloglin}
\alias{print.summary.genloglin}
\alias{print.predict.genloglin}
\title{Control Printed Display of Objects of Class "MMI"}
\usage{
\method{print}{MMI}(x, ...)

\method{print}{SPMI}(x, ...)

\method{print}{anova.genloglin}(x, ...)

\method{print}{genloglin}(x, digits = max(3, getOption("digits") - 3), ...)

\method{print}{predict.genloglin}(x, ...)

\method{print}{summary.genloglin}(
  x,
  digits = max(3, getOption("digits") - 3),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  ...
)
}
\arguments{
\item{x}{An object of class \code{"genloglin"}, produced by the
\code{\link{genloglin}} function.}

\item{...}{Additional arguments passed to or from other methods.}

\item{digits}{Minimum number of digits; see \code{\link{print.default}} for additional explanation.}

\item{symbolic.cor}{A logical value indicating whether correlations should be printed in symbolic form; see the \code{\link{summary}} method for class \code{\link{glm}} for additional explanation.}

\item{signif.stars}{A logical value indicating whether significance stars should be printed; see the \code{\link{summary}} method for class \code{\link{glm}} for additional explanation.}
}
\value{
A print out of selected results from \code{\link{MI.test}}.

A print out of selected results from \code{\link{MI.test}}.

A print out of selected results.
}
\description{
The \code{print.MMI} method function controls the printed display of objects
of class \code{MMI}.

The \code{print.SPMI} method function controls the printed display of
objects of class \code{SPMI}.

A function used to control the display of output provided by \code{anova.genloglin}.

Method functions that control the printed display of MRCV regression
modeling objects.

A function used to control the display of output provided by \code{predict.genloglin}.

A method function that controls the display of \code{summary.genloglin}
}
\details{
The \code{print.genloglin} function is based on the \code{\link{print}}
method for class \code{"glm"}.

The \code{print.summary.genloglin} function is based on the
\code{\link{print}} method for class \code{"summary.glm"}.
}
