\name{print.SPMI}
\alias{print.SPMI}
\title{
Control Printed Display of Objects of Class "SPMI"
}
\description{
The \code{print.SPMI} method function controls the printed display of objects of class \code{SPMI}.
}
\usage{
\method{print}{SPMI}(x, ...)
}
\arguments{
  \item{x}{
An object of class \code{"SPMI"} produced by the \code{\link{MI.test}} function.
}
  \item{...}{
Additional arguments passed to or from other methods.
}
}
\value{
A print out of selected results from \code{\link{MI.test}}.
}