\name{sealion}
\alias{sealion}
\docType{data}
\title{
Steller Sea Lion Scat Data of Riemer, Wright, and Brown (2011)
}
\description{
The sealion data frame is a portion of the data analyzed by Riemer, Wright, and Brown (2011).  The data represents the types of prey found in Steller sea lion scats found at the mouth of the Columbia river in August 2004 and 2007.  The goal was to determine if the diet habits of the sea lions had changed over time.
}
\usage{sealion}
\format{
The data frame contains the following 12 columns:

Column 1, labeled \code{Date}, corresponds to the date of data collection.  Two collection dates, 8/10/2004 and 8/7/2007, are included in the data frame.

Columns 2-12 correspond to the types of prey found in Steller sea lion scats.  Binary responses (1 = Present in scat, 0 = Not present in scat) are provided for each category.
\itemize{
\item{\code{prey1}: Unidentified fish}
\item{\code{prey2}: Pacific lamprey (\emph{Lampetra tridentata})}
\item{\code{prey3}: Starry flounder (\emph{Platichthys stellatus})}
\item{\code{prey4}: Pacific sardine (\emph{Sardinops sagax})}
\item{\code{prey5}: Pacific herring (\emph{Clupea pallasii})}
\item{\code{prey6}: Unidentified clupeid (family Clupeidae)}
\item{\code{prey7}: Unidentified skate (family Rajidae)}
\item{\code{prey8}: Northern anchovy (\emph{Engraulis mordax})}
\item{\code{prey9}: Pacific salmon (\emph{Oncorhynchus} spp.)}
\item{\code{prey10}: Pacific staghorn sculpin (\emph{Leptocottus armatus})}
\item{\code{prey11}: Pacific hake (\emph{Merluccius productus})}
}
}
\source{
Riemer, S. D., Wright, B. E., and Brown, R. F. (2011)  Food habits of Steller sea lions (\emph{Eumetopias jubatus}) off Oregon and northern California, 1986-2007.  \emph{Fishery Bulletin}, \bold{109}, 369--381.
}